move configdialog and build_config_layout into config.py.
fix pyside imports
add compatibility check for flags vs f in configdialog
