+++ examples/settings_dialog_demo.py
@@ -1,8 +1,10 @@
 import PyQt5
 from PyQt5.QtCore import Qt
 from PyQt5 import QtWidgets
-from math import ceil
-from pyqtconfig import ConfigManager
+# import PySide2
+# from PySide2.QtCore import Qt
+# from PySide2 import QtWidgets
+from pyqtconfig import ConfigManager, ConfigDialog
 
 default_settings = {
     "Setting 1": "Hello",
@@ -46,11 +48,7 @@ class MainWindow(QtWidgets.QMainWindow):
         self.settings_button.clicked.connect(self.create_config_dialog)
 
     def create_config_dialog(self):
-        config_copy = ConfigManager(self.config.defaults)
-        config_copy.set_many(self.config.as_dict())
-        config_copy.set_many_metadata(self.config.metadata_as_dict())
-
-        config_dialog = ConfigDialog(config_copy, self, cols=2, flags=Qt.WindowCloseButtonHint)
+        config_dialog = ConfigDialog(self.config, self, cols=2, f=Qt.WindowCloseButtonHint)
         config_dialog.setWindowTitle("Settings")
         config_dialog.setMaximumWidth(100)
         config_dialog.accepted.connect(lambda: self.update_config(config_dialog.config))
@@ -64,104 +62,6 @@ class MainWindow(QtWidgets.QMainWindow):
         self.current_config.setText(str(self.config.as_dict()))
 
 
-class ConfigDialog(QtWidgets.QDialog):
-    """
-    A Dialog class inheriting from QtWidgets.QDialog. This class creates layout from the input config using
-    build_config_layout, as well as QDialogButtonBox with Ok and Cancel buttons.
-    """
-    def __init__(self, config, *args, cols=None, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.config = config
-
-        # Build layout from settings
-        config_layout_kwargs = {} if cols is None else {"cols": cols}
-        config_layout = build_config_layout(config, **config_layout_kwargs)
-
-        # Create a button box for the dialog
-        button_box = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Reset | QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        button_box.accepted.connect(self.accept)
-        button_box.rejected.connect(self.reject)
-
-        # QDialogButtonBox places Reset after Ok and Cancel
-        button_box.buttons()[2].setText("Reset to Defaults")
-        button_box.buttons()[2].clicked.connect(self.show_confirm_reset_dialog)
-
-        # Place everything in a layout in the dialog
-        layout = QtWidgets.QVBoxLayout()
-        layout.addLayout(config_layout)
-        layout.addWidget(button_box)
-        self.setLayout(layout)
-
-    def show_confirm_reset_dialog(self):
-        message_box = QtWidgets.QMessageBox(text="Are you sure you want to reset to defaults?")
-        message_box.setWindowTitle("Warning")
-        message_box.setStandardButtons(QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Cancel)
-        message_box.buttonClicked.connect(
-            lambda button: (self.config.set_many(self.config.defaults) if button.text() == "OK" else None))
-        message_box.exec()
-
-
-def build_config_layout(config, cols=2):
-    """
-    Generate a layout based on the input ConfigManager. The layout consists of a user specified number of columns of
-    QFormLayout. In each row of the QFormLayout, the label is the config dict key, and the field is the config handler
-    for that key.
-
-    :param config: ConfigManager
-    :param cols: Number of columns to use
-    :return: QHBoxLayout
-    """
-    h_layout = QtWidgets.QHBoxLayout()
-    forms = [QtWidgets.QFormLayout() for _ in range(cols)]
-    for form in forms:
-        h_layout.addLayout(form)
-
-    num_prefer_hidden = sum([1 for _, value in config.metadata_as_dict().items() if value["prefer_hidden"]])
-    num_items = len(config.as_dict()) - num_prefer_hidden
-
-    i = -1
-    for key in config.as_dict():
-        if config.get_metadata(key)["prefer_hidden"]:
-            # If the key doesn't want to be shown, skip this row
-            continue
-        i += 1
-
-        # Find which column to put the setting in. Columns are filled equally, with remainder to the left. Each column
-        # is filled before proceeding to the next.
-        f_index = 0
-        for j in range(cols):
-            if (i+1) <= ceil((j+1)*num_items/cols):
-                f_index = j
-                break
-
-        # Get the handler widget for the key
-        if key in config.handlers:
-            # If we've already defined a handler, use that
-            input_widget = config.handlers[key]
-        elif config.get_metadata(key)["preferred_handler"] is not None:
-            # If there is a preferred handler in the metadata, create one of those. If there is a preferred mapper
-            # use that
-            input_widget = config.get_metadata(key)["preferred_handler"]()
-            if config.get_metadata(key)["preferred_map_dict"] is not None:
-                input_widget.addItems(config.get_metadata(key)["preferred_map_dict"].keys())
-                config.add_handler(key, input_widget, mapper=config.get_metadata(key)["preferred_map_dict"])
-            else:
-                config.add_handler(key, input_widget)
-        else:
-            # If there's no existing handler or preferred handler, try to create a default one. If it fails, skip this
-            # row
-            config.add_handler(key)
-            if key not in config.handlers:
-                continue
-            else:
-                input_widget = config.handlers[key]
-
-        label = QtWidgets.QLabel(key)
-        forms[f_index].addRow(label, input_widget)
-
-    return h_layout
-
-
 if __name__ == "__main__":
     app = QtWidgets.QApplication([])
     main_window = MainWindow()
@@ -169,4 +69,4 @@ if __name__ == "__main__":
     main_window.setMinimumSize(300, 250)
     main_window.show()
 
-    app.exec()
+    app.exec_()
